daemon: extract event/peer.rs from event/mod.rs
Move peer-config types and peer-group types out of mod.rs into a
dedicated event/peer.rs submodule:

- GrPeerConfig, RouteReflectorConfig, PeerConfig (static per-peer config)
- PeerParams and impl PeerParams (parameter builders for add/update peer)
- parse_afi_safis_yaml, parse_gr_yaml (YAML config parsers)
- impl TryFrom<&config::Neighbor> for PeerParams (YAML -> PeerParams)
- DynamicPeer, PeerGroup and impl PeerGroup (peer-group handling)

Types are re-exported from mod.rs as pub(super) so grpc.rs continues
to see them via its existing `use super::*` import without change.

PeerView stays in mod.rs because it holds both PeerConfig (from peer.rs)
and PeerState/MessageCounter (from mod.rs); moving it would create
circular imports.

No logic changes; all 937 tests pass unchanged.

mod.rs: 9,507 lines -> 8,974 lines
peer.rs: 561 lines (new)
