+++ daemon/src/event/mod.rs
@@ -152,6 +152,11 @@ impl MessageCounter {
 mod grpc;
 use grpc::GrpcService;
 
+mod peer;
+pub(super) use peer::{
+    DynamicPeer, GrPeerConfig, PeerConfig, PeerGroup, PeerParams, RouteReflectorConfig,
+};
+
 use crate::fsm::State as SessionState;
 
 /// Session-scoped counters and FSM state shared between `Peer` and the active
@@ -279,62 +284,6 @@ fn conn_effective_max(arb: &ConnArbiter, role: Option<crate::fsm::Role>, family:
         .and_then(|c| c.send_max().get(&family))
         .copied()
         .unwrap_or(1)
-}
-
-/// Static GR configuration for a single peer, set at peer creation.
-/// `None` in `PeerConfig::graceful_restart` means GR is disabled for this peer.
-/// Cloned into capability negotiation at each session open.
-#[derive(Clone)]
-pub(super) struct GrPeerConfig {
-    /// Restart Time advertised in our OPEN (12-bit, max 4095 s).
-    restart_time: u16,
-    /// Whether to set the N-bit (RFC 8538): GR applies to NOTIFICATION and Hold Timer expiry.
-    notification_enabled: bool,
-    /// Families included in the GR capability (non-empty by construction).
-    families: Vec<Family>,
-}
-
-/// RFC 4456 Route Reflector configuration for a single peer.
-#[derive(Clone, Default)]
-pub(super) struct RouteReflectorConfig {
-    route_reflector_client: bool,
-    /// Per-peer cluster ID; None means fall back to the local router-id.
-    route_reflector_cluster_id: Option<Ipv4Addr>,
-}
-
-/// Static per-peer configuration.  Set at peer creation (via `PeerParams::build`)
-/// and immutable for the lifetime of the peer.  Cloned into `PeerSession` at
-/// session start so the session task can access it without the global lock.
-#[derive(Clone)]
-pub(super) struct PeerConfig {
-    remote_addr: IpAddr,
-    remote_port: u16,
-    /// Expected AS number from configuration; 0 means "accept any".
-    /// The actual negotiated ASN lives in PeerState.remote_asn (session-scoped).
-    expected_remote_asn: u32,
-    local_asn: u32,
-    passive: bool,
-    delete_on_disconnected: bool,
-    holdtime: u64,
-    connect_retry_time: u64,
-    local_cap: Vec<packet::Capability>,
-    route_server_client: bool,
-    route_reflector: RouteReflectorConfig,
-    /// Snapshot of the global router-id taken at peer creation.
-    /// Immutable for the lifetime of the peer; used for RR ORIGINATOR_ID loop
-    /// detection without holding the global lock.
-    local_router_id: Ipv4Addr,
-    multihop_ttl: Option<u8>,
-    /// GTSM minimum TTL (RFC 5082); None = GTSM disabled.
-    /// When set, outgoing TTL is 255 and incoming packets below this value are dropped.
-    /// Takes priority over multihop_ttl when both are configured.
-    ttl_security: Option<u8>,
-    password: Option<String>,
-    /// Per-family prefix limits from config.
-    /// Used to initialize PeerSession::prefix_counters in accept_connection().
-    prefix_limits: FnvHashMap<Family, u32>,
-    /// GR helper config; None = GR disabled.
-    graceful_restart: Option<GrPeerConfig>,
 }
 
 /// Cross-session mutable state for a peer.
@@ -607,484 +556,6 @@ impl Peer {
                 .copied()
                 .unwrap_or(1),
         )
-    }
-}
-
-/// Plain-struct replacement for the old PeerBuilder.
-///
-/// Both TryFrom<&api::Peer> and TryFrom<&config::Neighbor> construct this as
-/// an exhaustive struct literal so that adding a new field causes a compile
-/// error at every construction site.
-pub(super) struct PeerParams {
-    remote_addr: IpAddr,
-    remote_port: u16,
-    expected_remote_asn: u32,
-    local_asn: u32,
-    passive: bool,
-    rs_client: bool,
-    route_reflector: RouteReflectorConfig,
-    delete_on_disconnected: bool,
-    admin_down: bool,
-    state: SessionState,
-    holdtime: u64,
-    connect_retry_time: u64,
-    multihop_ttl: Option<u8>,
-    ttl_security: Option<u8>,
-    password: Option<String>,
-    /// Per-family add-path mode (RFC 7911 2-bit flags); mode 0 means plain MP.
-    families: FnvHashMap<Family, u8>,
-    send_max: FnvHashMap<Family, usize>,
-    prefix_limits: FnvHashMap<Family, u32>,
-    graceful_restart: Option<GrPeerConfig>,
-}
-
-impl PeerParams {
-    const DEFAULT_HOLD_TIME: u64 = 180;
-    const DEFAULT_CONNECT_RETRY_TIME: u64 = 3;
-
-    /// Derive the local capability list from peer configuration.
-    ///
-    /// Separated from `build()` so the capability logic can be tested
-    /// independently of struct construction.
-    fn build_local_cap(
-        remote_addr: IpAddr,
-        local_asn: u32,
-        families: &FnvHashMap<Family, u8>,
-        graceful_restart: Option<&GrPeerConfig>,
-    ) -> Vec<packet::Capability> {
-        let mut local_cap: Vec<packet::Capability> = Vec::new();
-        if families.is_empty() {
-            local_cap.push(match remote_addr {
-                IpAddr::V4(_) => packet::Capability::MultiProtocol(Family::IPV4),
-                IpAddr::V6(_) => packet::Capability::MultiProtocol(Family::IPV6),
-            });
-        } else {
-            let mut addpath = Vec::new();
-            for (f, mode) in families {
-                if *mode > 0 {
-                    addpath.push((*f, *mode));
-                }
-                local_cap.push(packet::Capability::MultiProtocol(*f));
-            }
-            if !addpath.is_empty() {
-                local_cap.push(packet::Capability::AddPath(addpath));
-            }
-            // RFC 8950: advertise ExtendedNexthop when peering over IPv6
-            // with IPv4 address family configured.
-            // SR Policy (SAFI 73) is excluded: its nexthop is always the
-            // originator address in the same AFI, not an IPv6-mapped address.
-            if matches!(remote_addr, IpAddr::V6(_)) {
-                let enh_families: Vec<(Family, u16)> = families
-                    .keys()
-                    .filter(|f| f.afi() == Family::AFI_IP && **f != Family::IPV4_SRPOLICY)
-                    .map(|f| (*f, Family::AFI_IP6))
-                    .collect();
-                if !enh_families.is_empty() {
-                    local_cap.push(packet::Capability::ExtendedNexthop(enh_families));
-                }
-            }
-        }
-        if let Some(gr) = graceful_restart {
-            // N-bit (0x4): supports GR for NOTIFICATION and Hold Timer (RFC 8538).
-            // R-bit (0x8) is NOT set here; it is applied at connection time in
-            // PeerFsm::on_connected() based on the current global restarting state.
-            let flags = if gr.notification_enabled { 0x4 } else { 0 };
-            local_cap.push(packet::Capability::GracefulRestart {
-                flags,
-                restart_time: gr.restart_time,
-                families: gr.families.iter().map(|f| (*f, 0)).collect(),
-            });
-        }
-
-        // Always advertise 4-byte ASN support.
-        let four_octet = packet::Capability::FourOctetAsNumber(local_asn);
-        let four_octet_code: u8 = (&four_octet).into();
-        if !local_cap
-            .iter()
-            .any(|c| Into::<u8>::into(c) == four_octet_code)
-        {
-            local_cap.push(four_octet);
-        }
-        // Always advertise RFC 8654 Extended Message support.
-        local_cap.push(packet::Capability::ExtendedMessage);
-        local_cap
-    }
-
-    /// Build a `Peer` from these params.
-    ///
-    /// `local_router_id` is needed to construct `PeerFsm` for collision
-    /// detection; it is only known once `Global::router_id` is set, so callers
-    /// always go through `Global::add_peer` rather than calling this directly.
-    fn build(mut self, local_router_id: u32, global_asn: u32) -> Peer {
-        if self.local_asn == 0 {
-            self.local_asn = global_asn;
-        }
-
-        let local_cap = Self::build_local_cap(
-            self.remote_addr,
-            self.local_asn,
-            &self.families,
-            self.graceful_restart.as_ref(),
-        );
-
-        let conn_arbiter = Arc::new(std::sync::Mutex::new(ConnArbiter::new(
-            crate::fsm::PeerFsm::new(
-                local_router_id,
-                self.local_asn,
-                local_cap.clone(),
-                self.holdtime,
-                self.expected_remote_asn,
-                self.send_max.clone(),
-            ),
-        )));
-
-        Peer {
-            config: PeerConfig {
-                remote_addr: self.remote_addr,
-                remote_port: if self.remote_port != 0 {
-                    self.remote_port
-                } else {
-                    Global::BGP_PORT
-                },
-                expected_remote_asn: self.expected_remote_asn,
-                local_asn: self.local_asn,
-                passive: self.passive,
-                delete_on_disconnected: self.delete_on_disconnected,
-                holdtime: self.holdtime,
-                connect_retry_time: self.connect_retry_time,
-                local_cap,
-                route_server_client: self.rs_client,
-                route_reflector: self.route_reflector.clone(),
-                local_router_id: Ipv4Addr::from(local_router_id),
-                multihop_ttl: self.multihop_ttl,
-                ttl_security: self.ttl_security,
-                password: self.password,
-                prefix_limits: self.prefix_limits,
-                graceful_restart: self.graceful_restart,
-            },
-            admin_down: self.admin_down,
-            state: Arc::new(PeerState {
-                fsm: AtomicU8::new(self.state as u8),
-                peer_up_at: AtomicU64::new(0),
-                peer_down_at: AtomicU64::new(0),
-                remote_asn: AtomicU32::new(0),
-                remote_id: AtomicU32::new(0),
-                remote_holdtime: AtomicU16::new(0),
-                remote_cap: ArcSwapOption::empty(),
-                session_addrs: ArcSwapOption::empty(),
-            }),
-            counter_tx: Default::default(),
-            counter_rx: Default::default(),
-            context: Arc::new(std::sync::Mutex::new(PeerContext {
-                conn_arbiter,
-                active_connect_cancel_tx: None,
-                active_connect_join_handle: None,
-                gr_state: crate::gr::GrState::new(),
-                gr_restart_timer: None,
-            })),
-        }
-    }
-}
-
-/// Parse an afi-safis slice from YAML config into (families, send_max) maps.
-///
-/// Returns a map of Family -> add-path mode (RFC 7911 2-bit: bit0=RX, bit1=TX)
-/// and a separate send_max map for families where add-path TX is configured.
-pub(super) fn parse_afi_safis_yaml(
-    afi_safis: &[config::AfiSafi],
-) -> (FnvHashMap<Family, u8>, FnvHashMap<Family, usize>) {
-    let mut base_families: Vec<Family> = Vec::new();
-    let addpath_entries: Vec<(packet::Family, u8, usize)> = afi_safis
-        .iter()
-        .filter(|x| {
-            let name = x.config.as_ref().and_then(|c| c.afi_safi_name.as_ref());
-            let Some(f) = name else { return false };
-            if let Ok(family) = convert::family_from_config(f) {
-                base_families.push(family);
-            }
-            true
-        })
-        .filter_map(|x| {
-            let ap_config = x.add_paths.as_ref()?.config.as_ref()?;
-            let rx = ap_config.receive.unwrap_or(false);
-            let send_max = ap_config.send_max.unwrap_or(0) as usize;
-            let tx = send_max > 0;
-            let mode = u8::from(rx) | (u8::from(tx) << 1);
-            if mode == 0 {
-                return None;
-            }
-            let family =
-                convert::family_from_config(x.config.as_ref()?.afi_safi_name.as_ref()?).ok()?;
-            Some((family, mode, send_max))
-        })
-        .collect();
-
-    let mut families: FnvHashMap<Family, u8> =
-        base_families.into_iter().map(|f| (f, 0u8)).collect();
-    let mut send_max: FnvHashMap<Family, usize> = FnvHashMap::default();
-    for (f, mode, sm) in addpath_entries {
-        families.insert(f, mode & 0x3);
-        if sm > 0 {
-            send_max.insert(f, sm);
-        }
-    }
-    (families, send_max)
-}
-
-/// Build GrPeerConfig from YAML GracefulRestartConfig + per-family mp-graceful-restart flags.
-pub(super) fn parse_gr_yaml(
-    afi_safis: &[config::AfiSafi],
-    gr_config: Option<&config::GracefulRestartConfig>,
-) -> Option<GrPeerConfig> {
-    const DEFAULT_RESTART_TIME: u16 = 120;
-    if !gr_config.and_then(|c| c.enabled).unwrap_or(false) {
-        return None;
-    }
-    let gr_families: Vec<Family> = afi_safis
-        .iter()
-        .filter(|a| {
-            a.mp_graceful_restart
-                .as_ref()
-                .and_then(|gr| gr.config.as_ref())
-                .and_then(|c| c.enabled)
-                .unwrap_or(false)
-        })
-        .filter_map(|a| {
-            convert::family_from_config(a.config.as_ref()?.afi_safi_name.as_ref()?).ok()
-        })
-        .collect();
-    if gr_families.is_empty() {
-        return None;
-    }
-    Some(GrPeerConfig {
-        restart_time: gr_config
-            .and_then(|c| c.restart_time)
-            .unwrap_or(DEFAULT_RESTART_TIME),
-        notification_enabled: gr_config
-            .and_then(|c| c.notification_enabled)
-            .unwrap_or(false),
-        families: gr_families,
-    })
-}
-
-impl TryFrom<&config::Neighbor> for PeerParams {
-    type Error = String;
-
-    fn try_from(n: &config::Neighbor) -> Result<PeerParams, Self::Error> {
-        let c = n.config.as_ref().ok_or("missing neighbor config")?;
-        let afi_safis = n.afi_safis.as_deref().unwrap_or_default();
-
-        let remote_addr = c
-            .neighbor_address
-            .as_ref()
-            .ok_or("missing neighbor address")?;
-        let peer_as = c.peer_as.ok_or("missing peer-as")?;
-
-        let transport_config = n.transport.as_ref().and_then(|t| t.config.as_ref());
-        let timer_config = n.timers.as_ref().and_then(|t| t.config.as_ref());
-
-        let (families, send_max) = parse_afi_safis_yaml(afi_safis);
-        let graceful_restart = parse_gr_yaml(
-            afi_safis,
-            n.graceful_restart
-                .as_ref()
-                .and_then(|gr| gr.config.as_ref()),
-        );
-
-        // Extract per-family prefix limits.
-        let mut prefix_limits: FnvHashMap<Family, u32> = FnvHashMap::default();
-        for afi_safi in afi_safis {
-            let prefix_max = |pl: &Option<config::generate::PrefixLimit>| -> Option<u32> {
-                pl.as_ref()?.config.as_ref()?.max_prefixes
-            };
-            if let Some(v4) = &afi_safi.ipv4_unicast
-                && let Some(max) = prefix_max(&v4.prefix_limit)
-            {
-                prefix_limits.insert(packet::Family::IPV4, max);
-            }
-            if let Some(v6) = &afi_safi.ipv6_unicast
-                && let Some(max) = prefix_max(&v6.prefix_limit)
-            {
-                prefix_limits.insert(packet::Family::IPV6, max);
-            }
-        }
-
-        let holdtime = timer_config
-            .and_then(|c| c.hold_time)
-            .map(|v| v as u64)
-            .filter(|&v| v != 0)
-            .unwrap_or(PeerParams::DEFAULT_HOLD_TIME);
-        let connect_retry_time = timer_config
-            .and_then(|c| c.connect_retry)
-            .map(|v| v as u64)
-            .filter(|&v| v != 0)
-            .unwrap_or(PeerParams::DEFAULT_CONNECT_RETRY_TIME);
-
-        Ok(PeerParams {
-            remote_addr: *remote_addr,
-            remote_port: transport_config
-                .and_then(|t| t.remote_port)
-                .unwrap_or(Global::BGP_PORT),
-            expected_remote_asn: peer_as,
-            local_asn: c.local_as.unwrap_or(0),
-            passive: transport_config
-                .and_then(|t| t.passive_mode)
-                .unwrap_or(false),
-            rs_client: n
-                .route_server
-                .as_ref()
-                .and_then(|r| r.config.as_ref())
-                .and_then(|r| r.route_server_client)
-                .unwrap_or(false),
-            route_reflector: {
-                let rr_cfg = n.route_reflector.as_ref().and_then(|r| r.config.as_ref());
-                RouteReflectorConfig {
-                    route_reflector_client: rr_cfg
-                        .and_then(|r| r.route_reflector_client)
-                        .unwrap_or(false),
-                    route_reflector_cluster_id: rr_cfg
-                        .and_then(|r| r.route_reflector_cluster_id.as_deref())
-                        .filter(|s| !s.is_empty())
-                        .and_then(|s| Ipv4Addr::from_str(s).ok()),
-                }
-            },
-            delete_on_disconnected: false,
-            admin_down: c.admin_down.unwrap_or(false),
-            state: SessionState::Idle,
-            holdtime,
-            connect_retry_time,
-            multihop_ttl: n
-                .ebgp_multihop
-                .as_ref()
-                .and_then(|m| m.config.as_ref())
-                .and_then(|c| c.enabled.filter(|&en| en).and(c.multihop_ttl)),
-            ttl_security: n
-                .ttl_security
-                .as_ref()
-                .and_then(|ts| ts.config.as_ref())
-                .and_then(|c| {
-                    if c.enabled.unwrap_or(false) {
-                        Some(
-                            c.ttl_min
-                                .map(|v| if v == 0 { 255 } else { v })
-                                .unwrap_or(255),
-                        )
-                    } else {
-                        None
-                    }
-                }),
-            password: c.auth_password.clone(),
-            families,
-            send_max,
-            prefix_limits,
-            graceful_restart,
-        })
-    }
-}
-
-pub(super) struct DynamicPeer {
-    prefix: packet::IpNet,
-}
-
-pub(super) struct PeerGroup {
-    as_number: u32,
-    dynamic_peers: Vec<DynamicPeer>,
-    route_server_client: bool,
-    holdtime: Option<u64>,
-    local_asn: u32,
-    passive: bool,
-    route_reflector: RouteReflectorConfig,
-    multihop_ttl: Option<u8>,
-    ttl_security: Option<u8>,
-    auth_password: Option<String>,
-    connect_retry_time: Option<u64>,
-    families: FnvHashMap<Family, u8>,
-    send_max: FnvHashMap<Family, usize>,
-    graceful_restart: Option<GrPeerConfig>,
-}
-
-impl PeerGroup {
-    fn from_yaml(pg: &config::PeerGroup) -> Self {
-        let timer_config = pg.timers.as_ref().and_then(|t| t.config.as_ref());
-        let (families, send_max) =
-            parse_afi_safis_yaml(pg.afi_safis.as_deref().unwrap_or_default());
-        PeerGroup {
-            as_number: pg.config.as_ref().and_then(|c| c.peer_as).unwrap_or(0),
-            dynamic_peers: Vec::new(),
-            route_server_client: pg
-                .route_server
-                .as_ref()
-                .and_then(|rs| rs.config.as_ref())
-                .and_then(|c| c.route_server_client)
-                .unwrap_or(false),
-            holdtime: timer_config
-                .and_then(|c| c.hold_time)
-                .map(|h| h as u64)
-                .filter(|&h| h != 0),
-            local_asn: pg.config.as_ref().and_then(|c| c.local_as).unwrap_or(0),
-            passive: pg
-                .transport
-                .as_ref()
-                .and_then(|t| t.config.as_ref())
-                .and_then(|c| c.passive_mode)
-                .unwrap_or(false),
-            route_reflector: {
-                let rr = pg.route_reflector.as_ref().and_then(|r| r.config.as_ref());
-                RouteReflectorConfig {
-                    route_reflector_client: rr
-                        .and_then(|c| c.route_reflector_client)
-                        .unwrap_or(false),
-                    route_reflector_cluster_id: rr
-                        .and_then(|c| c.route_reflector_cluster_id.as_deref())
-                        .filter(|s| !s.is_empty())
-                        .and_then(|s| Ipv4Addr::from_str(s).ok()),
-                }
-            },
-            multihop_ttl: pg
-                .ebgp_multihop
-                .as_ref()
-                .and_then(|m| m.config.as_ref())
-                .and_then(|c| {
-                    if c.enabled.unwrap_or(false) {
-                        c.multihop_ttl
-                    } else {
-                        None
-                    }
-                }),
-            ttl_security: pg
-                .ttl_security
-                .as_ref()
-                .and_then(|ts| ts.config.as_ref())
-                .and_then(|c| {
-                    if c.enabled.unwrap_or(false) {
-                        Some(
-                            c.ttl_min
-                                .map(|v| if v == 0 { 255 } else { v })
-                                .unwrap_or(255),
-                        )
-                    } else {
-                        None
-                    }
-                }),
-            auth_password: pg
-                .config
-                .as_ref()
-                .and_then(|c| c.auth_password.as_deref())
-                .filter(|s| !s.is_empty())
-                .map(|s| s.to_string()),
-            connect_retry_time: timer_config
-                .and_then(|c| c.connect_retry)
-                .map(|t| t as u64)
-                .filter(|&t| t != 0),
-            families,
-            send_max,
-            graceful_restart: parse_gr_yaml(
-                pg.afi_safis.as_deref().unwrap_or_default(),
-                pg.graceful_restart
-                    .as_ref()
-                    .and_then(|gr| gr.config.as_ref()),
-            ),
-        }
     }
 }
